refactor: mybatis -> data jpa, somthing else
