--- .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Sun Mar 14 04:05:41 CST 2021
-gradle.version=6.7
