+++ build.gradle
@@ -1,34 +1,31 @@
 plugins {
     id 'java'
+    id 'org.springframework.boot' version '2.4.4'
+    id 'io.spring.dependency-management' version '1.0.8.RELEASE'
+    id 'io.freefair.lombok' version '5.3.0'
 }
 
 group 'org.example'
 version '1.0-SNAPSHOT'
 
 repositories {
-    maven {
-        url 'https://maven.aliyun.com/repository/public/'
-    }
-    mavenLocal()
-    mavenCentral()
+    maven { url 'https://maven.aliyun.com/repository/public/'}
 }
 tasks.withType(JavaCompile) {
     options.encoding = "UTF-8"
 }
 dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.6.0'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine'
-    testImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: '2.4.3'
+    testImplementation 'junit:junit:4.13'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
 
-    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-web', version: '2.4.4'
-    implementation group: 'com.alibaba', name: 'fastjson', version: '1.2.75'
-    implementation group: 'org.mybatis.spring.boot', name: 'mybatis-spring-boot-starter', version: '2.1.3'
-    implementation group: 'mysql', name: 'mysql-connector-java', version: '8.0.22'
-    implementation group: 'org.mybatis.spring.boot', name: 'mybatis-spring-boot-starter', version: '2.1.3'
-    implementation group: 'org.springframework', name: 'spring-jdbc', version: '5.3.4'
+    //mysql
+    implementation 'mysql:mysql-connector-java:8.0.22'
+    //H2Database
+    implementation 'com.h2database:h2:1.4.200'
+    //lombok
+    implementation 'org.projectlombok:lombok:1.18.16'
 
-}
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 
-test {
-    useJUnitPlatform()
 }
