--- src/main/java/recruit/company/config/CompanyConfig.java
@@ -1,12 +1,0 @@
-package recruit.company.config;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * @author w
- */
-@Configuration
-@ComponentScan(basePackages = {"recruit.company.controller", "recruit.company.service"})
-public class CompanyConfig {
-}
