--- src/main/java/recruit/company/entity/CandidateEntity.java
@@ -1,63 +1,0 @@
-package recruit.company.entity;
-
-import java.util.Date;
-
-/**
- * @author w
- */
-public class CandidateEntity {
-    private int id;
-    private String name;
-    private String state;
-    private String school;
-    private String phone;
-    private Date date;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getState() {
-        return state;
-    }
-
-    public void setState(String state) {
-        this.state = state;
-    }
-
-    public String getSchool() {
-        return school;
-    }
-
-    public void setSchool(String school) {
-        this.school = school;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    public Date getDate() {
-        return date;
-    }
-
-    public void setDate(Date date) {
-        this.date = date;
-    }
-}
