--- src/main/java/recruit/company/service/CandidateService.java
@@ -1,7 +1,0 @@
-package recruit.company.service;
-
-/**
- * @author w
- */
-public class CandidateService {
-}
