--- src/main/java/recruit/publicpac/config/PublicConfig.java
@@ -1,12 +1,0 @@
-package recruit.publicpac.config;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * @author w
- */
-@Configuration
-@ComponentScan(basePackages = {"recruit.publicpac.controller"})
-public class PublicConfig {
-}
