+++ src/main/java/recruit/service/CandidateService.java
@@ -1,0 +1,7 @@
+package recruit.service;
+
+/**
+ * @author w
+ */
+public class CandidateService {
+}
