+++ src/main/resources/application-dev.yml
@@ -1,0 +1,13 @@
+spring:
+  jpa:
+    database: h2
+  datasource:
+    driver-class-name: org.h2.Driver
+    username: root
+    password: root
+    url: jdbc:h2:~/employees
+  #    data: classpath:data-mysql.sql
+  #    initialization-mode: always
+  h2:
+    console:
+      enabled: true
