+++ src/main/resources/application-prod.yml
@@ -1,0 +1,11 @@
+spring:
+  jpa:
+    database: mysql
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.MySQL8Dialect
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    username: wuxc
+    password: hellowuxc
+    url: jdbc:mysql://localhost:3306/employees?serverTimezone=Asia/Shanghai
