+++ src/main/resources/application.yml
@@ -1,11 +1,25 @@
 server:
   port: 8010
 spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/recruit?serverTimezone=UTC&useUnicode=true&characterEncoding=utf-8
-    username: root
-    password: 123456
-
-mybatis:
-  typeAliasesPackage: recruit.dao
+  data:
+    jpa:
+      repositories:
+        enabled: true
+  jpa:
+    generate-ddl: true
+    hibernate:
+      ddl-auto: create-drop
+    show-sql: true
+    properties:
+      hibernate:
+        format_sql: true
+  jackson:
+    serialization:
+      FAIL_ON_EMPTY_BEANS: false
+    date-format: yyyy/MM/dd
+  main:
+    banner-mode: off
+  profiles:
+    active: dev
+  mvc:
+    throw-exception-if-no-handler-found: true
